+++ js/browser/bluebird.js
@@ -1,5 +1,5 @@
 /**
- * bluebird build version 2.4.1
+ * bluebird build version 2.4.2
  * Features enabled: core, race, call_get, generators, map, nodeify, promisify, props, reduce, settle, some, progress, cancel, using, filter, any, each, timers
 */
 /**
@@ -187,8 +187,8 @@ Async.prototype._consumeLateBuffer = function () {
 
 Async.prototype._queueTick = function () {
     if (!this._isTickUsed) {
-        this._schedule(this.consumeFunctionBuffer);
         this._isTickUsed = true;
+        this._schedule(this.consumeFunctionBuffer);
     }
 };
 
@@ -2219,8 +2219,7 @@ Promise.prototype.done = function (didFulfill, didReject, didProgress) {
 };
 
 Promise.prototype.spread = function (didFulfill, didReject) {
-    return this._then(didFulfill, didReject, undefined,
-        APPLY, undefined);
+    return this.all()._then(didFulfill, didReject, undefined, APPLY, undefined);
 };
 
 Promise.prototype.isCancellable = function () {
@@ -2632,36 +2631,13 @@ Promise.prototype._resolveFromResolver = function (resolver) {
     }
 };
 
-Promise.prototype._spreadSlowCase =
-function (targetFn, promise, values, boundTo) {
-    var promiseForAll = new PromiseArray(values).promise();
-    var promise2 = promiseForAll._then(function() {
-        return targetFn.apply(boundTo, arguments);
-    }, undefined, undefined, APPLY, undefined);
-    promise._follow(promise2);
-};
-
-Promise.prototype._callSpread = function (handler, promise, value) {
-    var boundTo = this._boundTo;
-    if (isArray(value)) {
-        for (var i = 0, len = value.length; i < len; ++i) {
-            if (tryConvertToPromise(value[i], promise) instanceof Promise) {
-                this._spreadSlowCase(handler, promise, value, boundTo);
-                return;
-            }
-        }
-    }
-    promise._pushContext();
-    return tryCatchApply(handler, value, boundTo);
-};
-
 Promise.prototype._callHandler = function (
     handler, receiver, promise, value) {
     var x;
+    promise._pushContext();
     if (receiver === APPLY && !this.isRejected()) {
-        x = this._callSpread(handler, promise, value);
+        x = tryCatchApply(handler, value, this._boundTo);
     } else {
-        promise._pushContext();
         x = tryCatch1(handler, receiver, value);
     }
     promise._popContext();